Implemented Discord OAuth2

+++ discord/discord-oauth2/src/main/java/net/drapuria/framework/discord/oauth2/entity/OAuth2Guild.java
@@ -19,7 +19,7 @@ public interface OAuth2Guild {
 
     String getIconUrl();
 
-    int getPermissionsRaw();
+    long getPermissionsRaw();
 
     EnumSet<Permission> getPermissions();
